diurnal diagnostics are now also work reasonably well if DTsrc is not 1hr.
The problem of the first few hours being not exactly correct(ADIURN) or missing
(HDIURN) for fields accumulated only when the radiation is called remains.

The diurnal vert.velocity cycle was eliminated (in ADIURN and HDIURN)

Potential out-of-bounds references for HDIURN were eliminated, and HDIURN
was reduced from (HR_IN_MONTH+4,.,.) to (HR_IN_MONTH,.,.). (Fortunately,
the previously existing out-of-bound references stayed within HDIURN !)

By the way, if DTsrc is a fraction of an hour, ADIURN/HDIURN collect the
mean over each hour. (Hopefully, we'll never try multiples of an hour).