Small addition to the dynamic documentation script. The script now also handles the variables in "USE ..., only: " statements.
The script provides only the names of the variables and the module
they are "USEd" from. One has to go to the module for description.
THe main reason I included these lists is that they help to distinguish
which variables are included in the list but are not actually used.
Such variables have red color.

Another convenient way to use this feature is to go to the list of ALL
variables. There for each global variable one can see all the places
(i.e. modules and subroutins) where it is used.

And just a reminder. You can always access the up-to-date dynamic
documentation for the code in repository at
 http://simplex.giss.nasa.gov/modelE/doc/
This documantation is updated automatically every day at 6am.