A couple of minor fixes in DIAG_PRT and DOMAIN_DECOMP (no change to results).
A new routine is written "daily_RAD" which sets daily radiation stuff
(including well-mixed greenhouse gases. This functionality was in
RADIA but now is moved out so that all values are known regardless of
whether there is a RADIA call that time step.

A new variable 'co2ppm' is defined in RAD_COM which holds the current
value of CO2 (as seen by the radiation code). (Does this need a
default value in the common block?).