Added documentation for a slighly obscure option for db-parameter o3_yr:
If o3_yr=-1950, e.g., dec1950 and jan1950 O3 are used to find O3 for
Jan 1-15 and Dec 16-31, rather than dec1949/jan1950 or dec1950/jan1951.
So the minus-sign may be used to avoid an end-of-year discontinuity.
(This is NOT true for any of the other ..._yr parameters, only for O3)