Commented out the second copy of the namelist to avoid confusion. It didn't cause any problems so far, but some compilers may not like it.
Also replaced "&END" with "/". This is necessary for some compilers
which don't want to work with "old style" namelist (NAG on Mac).
It works ok on Ra. If it causes any problems elsewhere please
rename it back to &END and let me know - I'll modify setup to
set appropriate ending depending on architecture.