This is an attempt to separate SURF_ALBEDO module completely from RADIATION. This is needed for future development since albedo is expected to become a non-trivial function of dynamic vegetation. As a result RADIATION.f was split into three files:    ALBEDO.f   - SURF_ALBEDO module    RADIATION.f - RADPAR module and programs that USE it    RAD_UTILS.f - everything else from RADIATION.f The creation of RAD_UTILS.f was necessary because some subroutines from RADIATION.f were used by both SURF_ALBEDO and RADPAR, so that leaving them in RADIATION.f would create dependency loops. I decided to move to RAD_UTILS.f also all subroutines which do not USE RADPAR module, just to reduce the size of RADIATION.f.
The data is exchanged between RADPAR and SURF_ALBEDO through the
list of formal parameters during the function call. I decided
against the use of USE statements here since that would require
to move a lot of global data from RADPAR to SURF_ALBEDO while
these data really belongs to RADPAR. It is also nice to see the
actual interface as a list of formal parameters.