Fixed a bug in getting ozone data, which occurred if O3_yr>0 (transient runs and O3_yr<0 were ok): after passing December 15, O3 from the next year (O3_yr+1) was used cyclically until the next restart.
This made the results restart dependent and O3 was off by a year for
most (but not all) of the run.

The fix is to treat O3_yr>0 the same as O3_yr<0, i.e. using December
and January of year O3_yr for the period from Dec 15 to Jan 15.