Several changes to model scripts to allow compilation and running of the model with MPI support. "gmake setup" now accepts three new options   ESMF=YES  - compile the model with ESMF enabled   MPI=YES   - set up model with MPI support   NPROC=<number of processors> - set the numbers of processors         for the first hour (i.e. "setup") run actually one needs both ESMF=YES MPI=YES to compile the model with ESMF.
MPI=YES tells the gmake to use "prun" (or mpirun on SGI) to run
the code. The number of processors is still passed from runE
through MP_SET_NUMTHREADS for compatibility with older runs,
but it will be changed later.

pdE and runpmE don't yet have MPI support. Will do it later.