Added support for one extra vegetation type: "C4 grass". To preserve full compatibility with previos versions the fraction of "C4 grass" is read from a separate file. The fraction of "C4 grass" is restricted to the amount of normal grass in the model, i.e. after the data is read it is re-arranged as folows:
 fract_c4_grass = min( fract_c4_grass, fract_grass)
 fract_grass    = fract_grass - fract_c4_grass

To enable this option one has to set

 read_c4_grass=1

in the rundeck. And specify the file with C4 grass fractions (in %):

VEG_C4=c4grass_72x46.ij

There is also file c4grass_144x90.ij for higher resolution runs.
Whithout this option the results are identical to previous version.

I didn't yet test properly the model with this option enabled. The
setup runs ok, but more testing is needed.