Inserted deallocation statement for local arrays in TRIDIAG_2D_DIST.
Can we agree on *always* deallocating temporary arrays when they are
not needed any more? Though this kind of garbage collecting for
local arrays seems to be included into f95 standard, for some
older compilers this constitutes a memory leak (like for XLF f90).
Also, the code looks cleaner when allocated memory is explicitly
deallocated.