Some cleaning to Makefiles to make sure that free-form fortran files compile on all machines. Currently the rules are:
files *.f are compiles as fixed-form fortran with CPP instructions
files *.F90 are compiled as free-form fortran with CPP instructions

The objects are specified as usual in the rundeck. For each object
gmake will look first for .f file and if not found it will look
for .F90 file. For obvious reasons one should not create
 .f and .F90 files with the same base name.

I replaced FFLAGSF with more intuitive F90FLAGS.
F90FLAGS are basically FFLAGS, but conain an extra options which
tells compiler that the file is in free-form.

I also took the liberty of removind rules like  %.o: %.U
Hopefully nobody is planning to use those files any more ...