Correction to two places in the code that caused floating point exception when pdE was called.
First was due to the fact that global array was treated as distributed
(i.e. hallo was assumed to exist). I tried to correct it in the spirit
of surrounding ESMF code, but it is still rather messy. Can somebody
tell me what "call ESMF_BCAST(grid_MIC, it)" is doing and is it necessary
to call it at the places that I corrected?

Second was due to the fact that "io_seaice" seems not to be called when
diagnostics are computed (I don't even see the real*4 I/O functionality
in it). But the arrays read in "io_seaice" are actually used during
the diagnostics calculations and cause floating point exception.
I initialized these arrays to 0.d0 which is probably not correct
and is ugly anyway. Maybe somebody more familiar with this part of the
code can have a look and decide if these arrays are necessary for
diagnostics.