--> added two routines for packing integers. This seemed to work for me, but maybe you want be critical when using them:    module procedure IPACK_1D      ! (i)    module procedure IPACK_2D      ! (i,j) while Tom Clune didn't check these specifically, he said if it worked for me I could commit, as a larger-scale change to the process is forthcoming...
--> added optinal logical argument to WRITE_PARALLEL* subroutines, called CRIT
(for criterion). Thus you can control whether or not a statement is written by
adding a condition to the call. E.g.
call write_parallel('hello world',crit=A.eq.B)
If the crit argument is not included in the call, the condition will default
to the usual criterion: AM_I_ROOT( ).

[Part of a large update for chemistry tracers to use MPI and a large update of
 the stratospheric chemistry.]