Included "mpi_defs.h", changed names of MPI calls so that the same functions are spelled in the same way (with regard to upper/lower case) since I need this for re-definitions in mpi_defs.h to work. Can we agree on some standard for spelling the names from external libraries?
Also fixed a bug in some rather hacky way. Tom, can you check it?
The problem was that in Esmf_LArrayGather_IJ and Esmf_IArrayGather_IJ
one had

          global_array(I1:IN,J1:JN) =
     &         RESHAPE(var(displs(J):displs(I)-1),
     &         shape=(/size(local_array,1),
     &         size(local_array,2)/))

inside a loop over NPES. The sizes of local domains may be different,
but the size of local_array stays the same, so in some cases
the extent (displs(J):displs(I)-1) is not equal
size(local_array,1)*size(local_array,2).