Added "private" attribute to DIAG_COM module and declared each of its elements as "public". You ask what the difference? Well, as far as intel is concerned, the difference is huge. The model now compiles with intel 9.0.031. (Without these modifications the compiler would hang while compiling DIAG_PRT.f).
My guess is that if module is not declared as "private" the compiler
would do a lot of extra work, like descending into all USE'd modules
and from them into even deeper modules, while in "private" module
it would look only for the variables that declared "public".
We should eventually put "private" into all modules. This may
improve the speed and reliability of compilations with intel.
And it is, of course, structurally correct thing to do anyway.