some tracer code tidying.
Note. In USE statements and declarations, preprocessor directives are
only required if there is a compile issue. If it is simply a question
of USE-ing more already exisiting variables, don't bother with the
directive.

I will try and reduce any clutter I come across, but everyone should
try and do their bit. Thanks!