New version of RMS: works with a template sheet I.RMS. I.RMS replaces some data statements in the previous version. RMS.f works for any resolution and does not use modelE modules. Unlike Gary's version, RMS does not (yet) update an existing score sheet; some coding to do so was left in RMS.f  .
Command line arguments of RMS: (no args -> usage instructions)
   IMxJM runID JAN_ij_file JUN_ij_file
IMxJM may be omitted if it is 72x46 AND runID starts with a letter,
JUN_ij_file may be omitted if it starts with JUN (rather than ../ or /)

Output files are called RMS_runID and AMS_runID.

The template sheet starts with 2 header lines (ignored by RMS).
The remaining lines are either blank (mirrored in the output) or
contain 8 fields and a title. The 8 fields have to be listed in the
proper order, their width however is irrelevant. The title has to be
exactly as in the ij_files to avoid getting the wrong or no field.

The names of the obs.data files are constructed using the grid dims IM,JM
and the first 2 columns of the template sheet, fieldID and ObsSourceID:
  name=OBS/fieldID//IMXJM.ObsSourceID (e.g. OBS/PREC144X72.LEGATES)

Note to GISS users:
A copy of RMS is saved in the /giss/exec directory (on poncho/athena,
  it is called /u/exec/RMSE to avoid overwriting Gary's version).

2 templates are stored in /giss/OBS (/u/cmrun/OBS):
  I.RMS1 corresponds to the previous version of RMS (good only for 72x46)
  I.RMS2 corresponds to Gary's 9/2006 version of RMS (=I.RMS of repository)
The corresponding obs. data files may be found in OBS/RMS1 and OBS/RMS2

Before using RMS, copy I.RMS1 or I.RMS2 to $PWD/I.RMS
                  link the corresponding OBS/RMS? directory to $PWD/OBS
                  link the appropriate topography file to $PWD/TOPO
($PWD is your present working directory, link means "ln -s")

The creation of the OBS/RMSx directories was a clumsy way to deal with
the proliferation of naming conventions for these files - standardization
seems to be called for ! Also, some files in OBS/RMS2 are updated versions
of files in OBS and I did not want to overwrite any files in OBS.
(note that some fieldIDs changed from I.RMS1 to I.RMS2)
Some conventions replace 144X72 by 2HX2, 72X46 by 5X4 (or 4X5)

My suggested standard would be: PREC_144x72.LEGATES