1) moved INC. SW. TOA diag outside rad loop for better averaging
2) removed some hard coding for vertical coordinate in RAD_DRV.f, SURFACE
and GHY_DRV (in particular, COE now removed with it's many complications)

Note, Looking at the code, there are literally dozens of places where we still
have P*SIG or DSIG or something. This is not sensible and I will try to
rationalise this as far as I can. In code you are responsible for, please check
to see whether you shouldn't be using the arrays already computed through
CALC_AMPK (in ATMDYN.f), or whether some more flexible scheme is required.
The idea is to hide the details of how pressures are set vertically from any
of the physics routines that don't really need to know.