Removal of vertical coordinate dependent code (i.e. that uses SIGE,DSIG etc.) from most of the model.
I have put in a new routine calc_vart_amp in ATMDYN which should be called
to return pressure/mass variables. There are still a few odd spots where I haven't
worked out what to do yet, but that'll come.

In doing this, I found a couple of errors in GWDRAG and in the diagnostics - not
too serious though.

There are three places where someone else needs to look:
- In CLOUDS2 there are some odd pressure related calculations with lots of hard
  coding from Surabi. This should be cleaned up.
- In TRACERS_Special_Lerner there is some possibly erroneous coding in
  get_wofsy_gas_IC
- In RAD_DRV.f, for the radiative forcing calculations there is a commented out
call to calc_amp. This should work - but should be checked.