Split ent_GISSveg.f into a driver (ent_prescribed_drv.f) and physics (still ent_GISSveg.f) files. The driver basically contains only i,j loops and reading routines. All PFT-dependent formulas and datasets are in ent_GISSveg.f .
New file ent_ENTveg.f is added. Currently it is identical to
ent_GISSveg.f . This is the file that should be changed to
create a new PFT model. I.e. all functions inside it should be
re-written for the new model.

When Ent is compiled with

 make ent

It will use GISS model (files ent_GISSveg.f, ent_pft.f)

When it is compiled with

 make ent PFT_MODEL=ENT

It will use "ENT" model (files ent_ENTveg.f, ent_pft_ENT.f)

All names of subroutines GISS_* were changed to prescr_*, so that
they are model-independent. When one changes the model one just
sustitutes a different file (currently ent_GISSveg.f or ent_ENTveg.f)
with these routines.

Ent compiles and runs as before for GISS model. It doesn't compile
for PFT_MODEL=ENT since *ENT*.f files have to be made compatible
with each other and with the rest of the model.