Cleaned up a bit more. Changed water limitation function (Wlim) back to original formula (rather than prescribing it as 1).
PS  FYI: This compiles fine (and I've tested it separately), but I got compilation errors for 2 other files after doing a new checkout earlier today: biophysics.f and ent_prescribed_updates.f. In the former, it's because C_lab is used as a patch variable but isn't defined so in ent_types.f, and in the latter, there are 4 vars (hemi, laidata, albedodata, cropsdata) which have both the intent and pointer attributes, which is apparently not allowable.