This script will execute 1hr and 1day runs (with restart) and compare results to ensure strong reproducibility for OpenMP and MPI configurations.
For each rundeck and configuration, the following operations are performed:
	- cvs co
	- build
	- run 1hr
	- continue 1dy

For the serial configuration the aux files are also constructed so as
to use CMPE002 for comparison.  An error is put in the log file for
those tests which fail.  Due to OpenMP issues on palm, a separate
serial build using "-mp" is performed for that comparison.

A comparison is also made against previously saved serial results.  If
all internal consistency checks passed, the assumption is that the
change is intentional and the old results are overwritten.  A
notification is added to the log that the change has been detected.

Currently, the following rundecks are tested: E1M20, E1oM20, E1F20,
E001tr.

This initial version should be considered draft, but does appear to
function adequately on palm/explore.

Next steps:
	- Modify cron to submit script each night on palm
	- Determine which mailing list should receive the logfile
	- Porting to halem.

Apologies that this was done is bash - should have been done in PERL.