Added a new option for MPI distributions: SCALI . These are relevant only for Linux machines so far. To specify MPI distributions use
   MPIDISTR=<distribution>

on a command line or in your .modelErc. If not specified the default
settings will be used (which should work on palm and halem). On
discover you have to specify:

   MPIDISTR=SCALI

This eliminates the need to use *_discover files committed earlier
today.

Another option supported is MPIDISTR=LAM . But that is of interest
only to those who have it installed on their computer, like desktop
linux box (LAM MPI is included in RedHat distribution).