Made a lot of trivial changes needed for single column model. Mostly added dimensions of arrays for serial gather/scatter. Also made an assumption that the SCM cell is in general position, i.e. HAVE_NORTH_POLE=HAVE_SOUTH_POLE=HAVE_EQUATOR=.false.
It was not clear what to do with GLOBALSUMs. I assumed that in serial
mode we want to sum over J_STRT:J_STOP, i.e. for SCM  GLOBALSUM
will return the value at J_SCM.

I think this completes the modifications to DOMAIN_DECOMP necessary
for SCM. But there are a lot of places in the model which need to
be modified before the SCM will work properly.