Merged more cleanups for MPI configuration.
Someone more knoweledgable about the continuing development of ODIFF
might want to take a few moments and look at some of the loops
associated with the poles that are currently commented out.  The loops
use a "fused index" for I,J that has been replaced in the remainder of
the routine with conventional (I,J) indices.  (More-or-less necessary
for MPI treatment.)