Found an uninitialized value in ODIFF() routine that was _exposed_ by the MPI conversion.  TANP(JM) was accessed, but never computed.  On halem no differences result, but on discover, differences occur after the 1st time step.
I remain convinced that any bit differences between serial and MPI
configurations should be aggressively investigated.  In this instance, the
CMPE002 differences look like roundoff, but the difference in values in
TANP(JM) were rather large in the two configurations.   This problem
presumably was also affecting the OpenMP configuration, but there are many
plausible reasons why we may not be able to force agreement with that
configuration.  (This is not to say we should not aim for that goal.)