Added a set of communication finctions with common interfaces
  SEND_TO_J(grd_dum, arr, j_dest, tag)
  RECV_FROM_J(grd_dum, arr, j_src, tag)

These are similar to MPI_Send and MPI_Recv, but instead of a rank
of a thread they use index J as a source/destination address.
These are needed for I/O in Dynamic Vegetation module.
To implement these I introduced a lookup table lookup_pet(J) which
is filled during the initialization of application. Tom, if you think
that this doesn't fit into the logic of DOMAIN_DECOMP or if such
lookup table is already present somewhere in ESMF, please let me
know or feel free to fix this.