This is a hack to fix a problem recently discovered in variable lakes runs. Basically this is an initial conditions problem. When I created .ext files long time ago they were assumed to be used for different land masks. All ground hydrology data, including snow, was extended to non-earth cells. Now that we use these files for variable lakes runs this creates a problem since snow can't emerge from the lake when lake shrinks.
Fixed this by removing snow from the cells fearth==0 at the restart
of the run. This should be done actually only at the setup of a new
run, but at the moment setting it up to do it on every restart so
that if people run into this problem they can continue their runs
after upgrade.