Some perl scripts to support automated regression testing - initially on "discover".
- CommandEntry.pm is a perl class that manages running a command either
  locally or via batch.  Writes a semaphore to indicate when command
  completes.

- CommandPool.m is a perl class that manages a suite of commands with
  dependencies.  As commands are added to the pool, they can be set to depend
  upon the completion of other tasks.  In this way, say, running of compiled
  executables can wait for completion of the compilation.

- regressionTests.pl - ad hoc driver for the above two packages.  Currently
  executes MPI, OPENMP, and SERIAL configurations of E1M20, E1oM20, E001tr,
  and E1F20.  MPI tests push these up to JM/2 processors.

Unfortunately I really screwed up and deleted one more package just before
this commit.  (I'm optimistic the admins will be able to recover the file.)
That package "RegressionTools.pm" is a suite of perl routines that wrap shell
commands for the various bits of building and running modelE.   I will commit
that file when the admins recover it.