Simple re-arrangement of declarations fixes (or, rather hides) the problem with E1oM20.R and -mp.
No, I didn't find a bug. Actually after playing with the code for
a while I am pretty sure that it is an optimizer problem. BTW,
reducing optimization to O1 for this file also make this problem
to go away. It seems that compiler incorrectly optimizes the call
to local subroutine (i.e. subroutine in the same file).
Replacing this call by a call to external subroutine (in another
file) which calls the same limitq_lin (inside OCNDYN.f) also
fixes the problem. So the subroutine itself seems to be ok.

I just changed the placement of declarations (as a C programmer
I find it painful to see local variables declared globally :) ).
Apparently ifort also finds it uncomfortable :).