Got rid of openunits, closeunits. This functionality is now included into openunit, closeunit which are made f90 interfaces.  In
  call openunit( filename, iunit, qbin, qold )
  call closeunit( iunit )

filename, iunit are either scalars or 1d arrays. qbin, qold are
always scalars and are optional (i.e. if you have to open both
formatted and unformatted files this should be done with
separate openunit calls). Corresponding changes have been made
where openunits, closeunits were used.

The users are advised to check if they need to add qold=.true.
to the calls to avoid changing modification date of input files.

Also, keep in mind that openunit called for an array of files
doesn't guarantee that unit numbers come as continuous sequence
of integers. So you can't assume this in the code.