As part of generalizing the 2d surface non-interactive source swapping, taking some fairly general routines out of TRACERS_SPECIAL_Shindell and into TRACERS.f (since others have expressed interest in eventually using it for other tracers). This is a work-in-progress, though the swapping works for the chemistry tracers.
The motivation here is to not have endless #ifdef options in TRACERS_DRV,
e.g. for GEIA, EDGAR_HYDE, EDGAR1995, IIASA, etc. emissions sets. Instead
have a standard file format and units and just swap the files.

For full information on creating the file format and header
and using this method, contact me (gfaluvegi@giss.nasa.gov).

Very quick summary: These added routines are:

(1) num_srf_sources: which counts the number of files called in the
rundeck: trname_01, trname_02... trname_XX. And automatically fills in
ntsurfsrc(n). Note that if you don't have any files of that form,
ntsurfsrc(n) gets set to zero, so you should still be able to "hardcode"
ntsurfsrc(n) after the call to num_srf_sources, if you are not using
this new method. But it is currently called for all tracers, so if you
happen to have a file in your rundeck of that form: I apologize for the
crash :-)

(2) read_sfc_sources: the routine that will then read the files found by
the num_srf_sources routine. There is a slightly new file format for the
new emissions files: an extra 80 character header record. Otherwise same
as normal GISS-binary IJ files. [netCDF ability could be added in the
future, I guess -- sorry Susa].

(3) read_emis_header: called from read_sfc_sources and num_srf_sources to
read that new header record, which contains the meta-data in the
following 80-character format, e.g.:

freq=a name=Alkenes source=GEIA_Industrial res=M

where freq is one-character, either a for annual (expects one post-
header record) or m for monthly (expects 12 post-header records)

where name is the name of the tracer, and this should match
trim(trname(n)). Up to 10 characters.

where source is the name of the source in the file, and this will
be used to label all the diagnostics (e.g. lname and sname). Up to
50 characters.

where res is the model horizontal resolution (though nothing is done with this
at the moment.) Expects one character (e.g. M, F, C for the medium, fine
coarse definitions. e.g. M for 4x5 deg model).

(4) parse_header: this routine called from read_emis_header to get the
variables out of the header string.

(5) read_mon_src_2: Called from read_sfc_sources when sources are
monthly. This is the actual read routine that calls e.g. READT_PARALLEL.
There is a problem with this routine: I didn't want to have 2 big global
arrays saved in a "save" statement: e.g. the equivalent of Jean's
tcla and tclb which are used to keep monthly files open and so only read
them every half-month, I think. Because mine would have been 2 arrays of
size: (IM,JM,ntm,ntsurfsrcmax). So instead, the routine re-reads the
montly files daily. I am sure there is a better way to do this in the
future to be more efficient. If that gets figured out, we could just use
the existing read_monthly_sources.

Since this is a big update, please contact me if it causes problems with
your tracers.