I am taking a liberty of committing a different version of cubic equation for "ci" in FBBphotosynthesis::ci_cubic . By default all new changes are commented out with #ifdef statements so the program still produces exactly the same results as before. To try new solution uncomment the line
!#define USE_IGORS_CUBIC

in  ci_cubic. Also if you want to see some debugging information
(both for old an new versions) uncomment the line

!#define DEBUG_CUBIC

I am claiming that the new version produces a correct solution to
the system of equations we are solving while old version gives
different A's if ci is substituted in

      A_1 = (cs_ - ci)/(C6*rs_)
      A_2 = (ca-cs_)/(C3_*rb)

I also took the liberty of changing interface of cubicroot(a,b,c,d,x,n)
as I think it is supposed to be (i.e. it should return ALL roots and
not just one). "x" is the array of roots, "n" is number of roots.

The idea of taking the maximal root seems to be almost correct one.
What one really wants is to get a positive "A" and from the output I
see there is either one A>0 (corresponding to maximal ci) or none
at all. So in most cases this approach seems to work. But for the
case when all A's are negative, I think, some extra work needs to
be done. My guess is that formula

    rs_ = 1./(m*A_*rh/cs_ + b)

is not applicable in case A<0 . You probably want to use abs(A)
in this case or just set rs=1/b ...