Finally got around to fixing the aliased pointers among SD, PIT, and CONV in the DYNAMICS module.  Originally these variables were EQUIVALENCEd, but dynamic allocation forced us to use pointer association instead, which unfortunately resulted in a shift in the J lower bounds when using SD and PIT in expressions.
SD, PIT, and CONV are now completely separate arrays.   This does require
an additional copy back into CONV which is used external to dynamics to
fill SD_CLOUDS.   That copy could be directly from PIT and SD, but perhaps
that is not as clear?

The best part of this is the elimination of the "JJ" macro at the top of a small
number of files.