Changed the way the makefile finds which .mod files are created after each compilation command. Now it uses .depend* file for this purpose (previously used a time stamp).
This makes the Makefile multithread-compatible, i.e. make can run
on parallel processors (if more than one are available). All processors
should be on the same node of course. So if you are in "batch shell"
mode on discover you can try to compile with

   make -j 4 setup ...

this will run up to 4 compilation jobs at the same time when possible.
I didn't check though if it really speeds the things up.

All this makes no difference if compiling on one processor.