Added a new function "ci_cubic1" which incorporates the latest solution to Ball-Berry/Farquhar cubic equation. Seems to work fine but it was not yet well tested.
To use it rename "ci_cubic" to something different (like ..._old)
and rename "ci_cubic1" to "ci_cubic". Or insert it at the end
of "ci_cubic" to compare printouts.

For A<0 this solution falls on 1/rs = b , in which case it
solves quadratic. One (undesirable?) "feature" of this solution is
that ci may become very big (up to 4000 in my tests) . This is a result of
very high rs (since b is very small). For comparison, typical
values of ci for A>0 are 200 - 250 .