Update of new downdraft/evap code to incorporate tracers properly - so far only for TRACERS_WATER though.
In doing this, I fixed two errors in the main model (one that meant that
qprime was not updated to the last iteration and that the skin-temperature
was not using the correct wind speed). These are unlikely to make much
difference but will change results.

I have also started to tidy up SURFACE.f which was(is) a mess. I have also
tried compacting some unnecessarily verbose tracer and diag code - hopefully
without messing anything up. If you notice anything strange, let me know.