Added an option to generate a netcdf template file from an ascii template provided by user:
If decks directory contains a file <run_id>.nctemp then during setup
this file will be copied to run directory with the name "nctemp".
On each restart (setup or continuation) if this file is present
"ncgen" will be called:

  ncgen -b -o nctemp.nc nctemp

i.e. corresponding binary file "nctemp.nc" will be generated.

For the path to "ncgen" I use $NECDFHOME/bin which seems to work on
discover, athena and with default linux distribution.