Beta version of utilities to facilitate output of model quantities in netcdf format at arbitrary time resolution.
To output a new quantity, two steps are required:

(1) Declaration of this quantity and its dimensions etc.
in a netcdf template file: decks/RUN_NAME.nctemp -
a self-documenting example is in model/netcdf_template.cdl

(2) A call to one of the inst_ncout* subroutines located
in INST_netcdf.f.  This can be done from anywhere in the
model having access to the (potentially user-defined)
variable in question.

To preclude any unforseeable disruptions to the normal functioning
of the model for the majority of users unlikely to ever employ this
capability,  "vanilla" instances of the procedure have not yet been
added to the main code.

The functioning of this system has only been lightly tested.
Behavior for extremely large output files and other such
situations is unknown.