First steps toward making the diagnostics compatible with different wind velocity grids. Printout results remain identical for now.
Numerous B-grid-specific averaging factors were moved
from DIAGJK and DEFACC/DIAG_PRT routines to DIAGA/DIAGB.

DIAGA/DIAGB are still B-grid specific.  Depending on interest
in the more esoteric diagnostics, DIAGA/DIAGB will either
(1) be preserved exactly as they are, to be optionally used
in conjunction with the B grid dynamics, or
(2) generalized to handle other grids in a way which does not
produce exactly the same results.

GEOM module now contains parameters defining the latitudinal
placement of U and KE in the dynamics, to be used by the output
routines.
For comparing different dynamics schemes, the freedom to have
wind-related diagnostics on a grid other than the dynamical
one has been left in the output routines (complete programming of
this TBD).

Horizontal/vertical grid information now passed to POUT_IJK in
POUT_netcdf.f.