A while back I added the functionality (for any tracer) for swapping emissions files in the rundeck, or adding them to sectors or regions, without any code alterations or re-compiling. Please see the cvs log, e.g. for TRACERS.f to notes on how to use that.
Now I am committing an extension (also available to all tracers by
using rundeck filenames of the form tracername_01, tracername_02, ... etc.) to have
transient emission files. The header now has an optional variable y=####-####.
For example if you have y=1890-1990, it will assume that you have 11 post-header
records (for annual sources, 12*11 for monthly) representative of 1890, 1900, 1910,
... 1990, and it will use the model jyear/jday to linearly interpolate to the present value.
If the start year and end year are equal (or if there is no y= in the header), the
source is treated as non-transient. If the model time is before the start year, it uses the
start year value, and if it is after the end year, it uses the end year value. (I believe
you could still use Jean's itime_tr0 to shut off emissions before a certain time?)

Source name in the header is now limited to 30 characters (was 50).

By the way: no spaces in header variables please.

PLEASE NOTE:
This is not heavily tested yet, but needs to be committed because someone needs
to start using it. Also, this is still not entirely ready, because I have the sources reading
files daily, i.e. much more often than they need to be read. I need to have a discussion
with a better coder as to the memory managemanet of saving the large
arrays needed to avoid this. Please contact me if it causes any trouble for you:
gfaluvegi@giss.nasa.gov

Committing changes to TRACERS.f, TRACER_COM.f (for general tracers).

Also committing new TRACERS_SPECIAL_Shindell.f and change to the sample rundeck
E001TdsM23.R for specific cases of 3D sources like aircraft and GFED biomass burning.)