Since other people started working on HYCOM parallelization I thought I'd commit my recent changes. The code may still be messy at some parts and it contains a lot of debugging stuff but it is usable. Those not working on HYCOM parallelization may want to postpone updating their HYCOM code at this moment.
The main change is that I switched from global "dimensions" arrays
(i.e. the ones in hycom_dim) to distributed arrays. So as with
hycom_arrays one is supposed to use "hycom_dim_glob" for serial
parts of the code and "hycom_dim" for parallelized code.
"hycom_dim" arrays are first initialized in parallel in bigrid.f
and then "gathered" to global arrays (available only on ROOT).
Only "ip" array has meaningful halos (couldn't update halos
in all arrays because we don't have halo_update for integers).

cnuity.f is pretty much parallelized. Probably still needs some
testing/optimizing.
bigrid.f is also running in parallel and is producing correct
"hycom_dim" arrays ("ip" with correct halos).

The rest of the code was pretty much hacked to allow parts of it
to run in parallel. In particular a lot of
  if (AM_I_ROOT()) then ... endif
instructions are included inside hycom.f to allow some of its parts
to run in parallel.

I introduced a PERIODIC_INDEX(x,y) macro in hycom_mpi_hacks.h .
This is for indices ja,jb which should be defined differently
in serial and parallel mode. Basically in the code one just
should use

 ja = PERIODIC_INDEX(j-1, jj)
 jb = PERIODIC_INDEX(j+1, jj)

see how they are used in cnuity.f . If you have suggestions on
a better way to implement this please let me know.

Currently the code runs both in serial and MPI-parallel mode
(checked up to 1 day run on 1,2,3 cpu's). Not tested with OpenMP.

E1fzhyc.R is the latest rundeck that should be used with HYCOM code.

DOMAIN_DECOMP.f : added IPACKj_2D to pack integer arrays.