Added a new rundeck option "reset_canopy_ic" (reset canopy initial conditions) which allows to reset canopy prognostic variables to some reasonable values.
If reset_canopy_ic=1 in the rundeck, the canopy temperature will be
reset to tearth(i,j) (from GIC) and canopy water (and water tracers
if enabled) will be set to 0. This option is of interest only to those
who want to change vegetation cover but want to start the model
from default GIC file. This option has effect only for istart < 9
(i.e. during the setup). The defaullt is 0 (i.e. do not reset
canopy variables).