--> Changing to what I believe are J.Lerner's latest parameters for the gravity wave / strat drag from E001TM23.R ?
--> added options (though turned off) for:
#define INTERACTIVE_WETLANDS_CH4, #define GFED_3D_BIOMASS,
#define BIOGENIC_EMISSIONS

--> moved some general tracer code before specific tracer code in the
order of the object code listings...

--> added SparseCommunicator_mod, in case needed.

--> Listed input files for interactive wetlands. Also listed the
parameters for it (note that ns_wet is important to altering the
correct source.)

--> removed commented input files for Koch tracers [which
reminds me, we are in need of a sample deck with more
tracers on, I think.]

--> added a vegetation sector to the sample emissions sector
setup, but this does nothing by default.

---> some slight clarification of the options for initial conditions
for CH4. Though still not clear.

--> removing the PIratio_indus and PIratio_bburn parameters, since
you should be using the sectors/regions method to alter sources now.